+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Added Craft 4.0 compatibility 
+
+### Changed
+- Changed plugin icon
+- Child Me! now requires Craft 3.7.0
+
 ## 1.2.0 - 2020-11-21  
 
 ### Added